+++ src/RuleSet/Php53.php
@@ -39,7 +39,6 @@ final class Php53
                 new Fixer\MultilineCommentOpeningClosingAloneFixer(),
                 new Fixer\NoDuplicatedArrayKeyFixer(),
                 new Fixer\NoDuplicatedImportsFixer(),
-                new Fixer\PhpdocTagNoNamedArgumentsFixer(),
                 new Fixer\PhpdocTypesCommaSpacesFixer(),
                 new Fixer\PhpUnitRequiresConstraintFixer(),
                 new PhpCsFixer\Fixer\Whitespace\LineBreakAfterStatementsFixer(),
@@ -58,10 +57,6 @@ final class Php53
                 ],
                 'PhpCsFixerCustomFixers/no_duplicated_imports' => true,
                 'PhpCsFixerCustomFixers/php_unit_requires_constraint' => true,
-                'PhpCsFixerCustomFixers/phpdoc_tag_no_named_arguments' => [
-                    'description' => '',
-                    'directory' => '',
-                ],
                 'PhpCsFixerCustomFixers/phpdoc_types_comma_spaces' => true,
                 'align_multiline_comment' => [
                     'comment_type' => 'all_multiline',