+++ test/Unit/RuleSet/Php80Test.php
@@ -49,7 +49,6 @@ final class Php80Test extends ExplicitRuleSetTestCase
             new Fixer\MultilineCommentOpeningClosingAloneFixer(),
             new Fixer\NoDuplicatedArrayKeyFixer(),
             new Fixer\NoDuplicatedImportsFixer(),
-            new Fixer\PhpdocTagNoNamedArgumentsFixer(),
             new Fixer\PhpdocTypesCommaSpacesFixer(),
             new Fixer\PhpUnitRequiresConstraintFixer(),
             new PhpCsFixer\Fixer\Whitespace\LineBreakAfterStatementsFixer(),
@@ -80,10 +79,6 @@ final class Php80Test extends ExplicitRuleSetTestCase
             ],
             'PhpCsFixerCustomFixers/no_duplicated_imports' => true,
             'PhpCsFixerCustomFixers/php_unit_requires_constraint' => true,
-            'PhpCsFixerCustomFixers/phpdoc_tag_no_named_arguments' => [
-                'description' => '',
-                'directory' => '',
-            ],
             'PhpCsFixerCustomFixers/phpdoc_types_comma_spaces' => true,
             'align_multiline_comment' => [
                 'comment_type' => 'all_multiline',