+++ CHANGELOG.md
@@ -12,6 +12,10 @@ For a full diff see [`6.49.0...main`][6.49.0...main].
 
 - Updated `friendsofphp/php-cs-fixer` ([#1249]), by [@dependabot]
 
+### Fixed
+
+- Disabled the `PhpCsFixerCustomFixers/phpdoc_tag_no_named_arguments` fixer ([#1251]), by [@localheinz]
+
 ## [`6.49.0`][6.49.0]
 
 For a full diff see [`6.48.0...6.49.0`][6.48.0...6.49.0].
@@ -1995,6 +1999,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#1244]: https://github.com/ergebnis/php-cs-fixer-config/pull/1244
 [#1245]: https://github.com/ergebnis/php-cs-fixer-config/pull/1245
 [#1249]: https://github.com/ergebnis/php-cs-fixer-config/pull/1249
+[#1251]: https://github.com/ergebnis/php-cs-fixer-config/pull/1251
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler